backend/requirements.txt
Flask
Flask-Cors
opencv-python
ultralytics
numpy
Pillow
